requirements.txt
anyio==4.11.0
annotated-types==0.6.0
boltons==25.0.0
certifi==2025.10.5
cffi==1.17.1
charset-normalizer==3.3.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
filelock==3.19.1
fonttools==4.60.1
h11==0.16.0
httptools==0.7.1
Jinja2==3.1.6
kiwisolver==1.4.9
matplotlib==3.10.7
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pillow==11.3.0
polars==1.35.1
psutil==7.1.3
pycparser==2.21
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.6.1
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.3
requests==2.32.4
rich==13.9.4
scipy==1.16.3
setuptools==78.1.1
six==1.17.0
sympy==1.14.0
torch==2.7.1+cu118
torchaudio==2.7.1+cu118
torchvision==0.22.1+cu118
tqdm==4.67.1
triton==3.3.1
ultralytics==8.3.225
ultralytics-thop==2.0.18
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
wheel==0.45.1

# FastAPI and file upload support
fastapi==0.101.1
python-multipart==0.0.6
